pyproject.toml
[tool.poetry]
name = "system-iii"
version = "0.1.0"
description = ""
authors = ["Ivan Zatevakhin"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.0"
ollama = "^0.4.0"
langchain-ollama = "^0.2.0"
pymumble = {git = "https://github.com/zatevakhin/pymumble.git", rev = "feat/build-from-proto"}
resampy = "^0.4.3"
rich = "^13.7.1"
click = "^8.1.0"
reactivex = "^4.0.4"
langchain-core = "^0.3.0"
langchain-community = "^0.3.0"
pysilero-vad = "^2.0.1"
watchdog= "^6.0.0"
soundfile = "^0.13.0"
qdrant-client = "^1.13.0"
"pyannote.audio" = "^3.3.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.0"
pytest = "^8.3.0"
ipykernel = "^6.29.4"
jupyterlab = "^4"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

line-length = 120

exclude = [
    ".git",
    ".venv",
    ".devenv",
    ".direnv",
]

fixable = ["ALL"]
unfixable = ["F401"]

[tool.ruff.lint.isort]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
